(#1289) - Prevent WebSQL->IDB data loss
If a browser is using both WebSQL and IDB,
and if a new database is created without
specifying the adapter, check if a DB
with the same name exists in WebSQL before
choosing IDB.

This ensures that no data loss occurs
if a browser upgrades from WebSQL-only
to WebSQL+IDB (e.g. Android 4.4).

This also adds some additional browser
migration tests.